Update README.md
I am starting to see a bit of the ambiguity in the defensive stats.  For example, it seems to me that FL should be FF (fumbles forced) if we are looking at this from the first person perspective of a defense.  I mean perhaps that is just semantics, but it seems like when the spreadsheet itself is alternating perspectives, the variables names would be important.